Add files via upload
First project. Started learning coding about 3 weeks, and bought a course on udemy. Wanted to test my "knowledge", so I picked what seemed to be a simple and straight forward project on Frontend Mentor. And yea, soon as i opened the index.html I was lost. Completed it in about 2 days with all the resources and documents I could find. The code is a mess, its all a mess. I am hoping to get better as long I keep trying..